Fix static
